Overhaul metadata, remove migrated methods.
